Conflicts:
	mongodb/repl/src/main/java/com/torodb/mongodb/repl/guice/MongoDbReplModule.java

+++ mongodb/repl/src/main/java/com/torodb/mongodb/repl/commands/impl/DropDatabaseReplImpl.java
@@ -18,15 +18,16 @@
  * 
  */
 
-package com.torodb.mongodb.repl.commands;
+package com.torodb.mongodb.repl.commands.impl;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import com.eightkdata.mongowp.Status;
 import com.eightkdata.mongowp.server.api.Command;
 import com.eightkdata.mongowp.server.api.Request;
 import com.eightkdata.mongowp.server.api.tools.Empty;
 import com.torodb.core.exceptions.user.UserException;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import com.torodb.torod.SharedWriteTorodTransaction;
 
 /**
@@ -43,7 +44,13 @@ public class DropDatabaseReplImpl extends ReplCommandImpl<Empty, Empty> {
         try {
             LOGGER.info("Dropping database {}", req.getDatabase());
 
-            trans.dropDatabase(req.getDatabase());
+            if (trans.existsDatabase(req.getDatabase())) {
+                trans.dropDatabase(req.getDatabase());
+            } else {
+                LOGGER.info("Trying to drop database " + req.getDatabase() + " but it has not been found. "
+                        + "This is normal since the database could have a collection being filtered "
+                        + "or we are reapplying oplog during a recovery. Ignoring operation");
+            }
         } catch (UserException ex) {
             reportErrorIgnored(LOGGER, command, ex);
         }
